fix: slider shrinks too much
