+++ package.json
@@ -29,7 +29,9 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.0",
+    "react-slick": "^0.30.2",
     "sass": "^1.78.0",
+    "slick-carousel": "^1.8.1",
     "zod": "^3.23.8"
   },
   "devDependencies": {
@@ -44,6 +46,7 @@
     "@types/node": "^22.5.4",
     "@types/react": "^18.3.5",
     "@types/react-dom": "^18.3.0",
+    "@types/react-slick": "^0.23.13",
     "eslint": "^9.10.0",
     "eslint-config-next": "14.2.9",
     "eslint-config-prettier": "^9.1.0",
