--- src/components/common/Slider/Slider.module.scss
@@ -1,45 +1,0 @@
-@import '@/theme/variables';
-@import '@/theme/mixins';
-
-.sliderContainer {
-  position: relative;
-  height: 780px;
-  width: 100%;
-  overflow: hidden;
-}
-
-.sliderImage {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  opacity: 0;
-  transition: opacity 0.5s ease-in-out;
-}
-
-.sliderImage.active {
-  opacity: 1;
-}
-
-.indicators {
-  position: absolute;
-  bottom: 20px;
-  left: 50%;
-  transform: translateX(-50%);
-  display: flex;
-  gap: 10px;
-}
-
-.indicator {
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  background-color: rgba(255, 255, 255, 0.5);
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.indicator.activeIndicator {
-  background-color: $white-text;
-}
