+++ src/features/home/sections/FirstViewSection/components/ImageSlider/ImageSlider.module.scss
@@ -1,0 +1,16 @@
+@import '@/theme/variables';
+@import '@/theme/mixins';
+
+.sliderContainer {
+  position: relative;
+  width: 100%;
+  // height: 780px;
+  height: auto;
+  overflow: hidden;
+
+  img {
+    width: 100%;
+    height: auto;
+    object-fit: cover;
+  }
+}
